requirements.txt
albumentations
open_clip_torch
ultralytics
dill
python-telegram-bot
requests